Adding linker, phase and transport representations.
NEW: The Linker class implements Linker that connect two phases in a model. Linkage manages those Linker and implements their implementation in a model.

NEW: The Transport class is an adaptation of cobra's Reaction class, which uses an adapted version of metabolites, allowing them to be assigned to a phase.

NEW: The phase class defines the individual states of the model. Based on these, the model is duplicated.

NEW:  Phases class contains multiple Phase objects and can apply them to a model.

UPDATED: Move the files necessary for duplicating the model to the duplication folder.