Simultaneous addition of linkers and individual models for phases.
UPDATED: GitHub workflow now has a test that tests the 'requirements.txt'. This is done using tox.

UPDATED: Phase now has a toString method.

UPDATED: Phases can now contain a model. For these phases, 'apply_phases' will not create a copy of the passed model, but will use the model associated with the phase. This means that a suffix is added to all elements of this model and otherwise the model is added to the result model like any other newly created model.

NEW: constraints.py has a new function 'add_linker_series'. Using this, linkers can be created across all time phases.