Visualization of groups using MetExploreViz.
NEW: Function that creates/downloads all the necessary data for the usage of MetExploreViz and starts a local web server to access MetExploreViz.

NEW: index.html has been added to the resources. This file is used for starting MetExploreViz.

NEW: Internal function that returns all associated metabolites, reactions and groups based on a COBRApy group.

UPDATED: The functions that convert COBRApy models to a MetExploreViz readable format now have an additional parameter that can be used to define groups. This ensures that only these groups are used and elements that do not belong to them are ignored during translations.