Integration of MetExploreViz using a custom widget.
NEW: Widget that lets you select groups of a COBRApy model and then visualize them using MetExploreViz.

NEW: Function that creates a multi-checkbox-widget.

UPDATED: Changes to the paths within the HTML file used for MetExploreViz.

UPDATED: The Jupyter Notebook for visualization now also shows the use of the new widget.

UPDATED: The phase visualization using Cytoscape uses the new widget if a model has been defined for the selected phase.