Selection of side compounds
NEW: New function 'select_side_metabolites' that creates the widget for the selection of side compounds.

UPDATED:  'metexplore_select_groups' has been renamed to 'metexplore_interface'.

UPDATED: 'metexplore_interface' now also supports the selection of side compounds as defined by MetExploreViz. 

UPDATED: cobra2metexplore' can now either create a JSON in which elements from not selected groups are marked as hidded but still included or are simply not included as before.

UPDATED: The index.html was extended so that now also the side compounds are loaded automatically.

BUGFIX: The flux values are now loaded correctly. Previously, a mismatch between identifiers in the file and within MetExploreViz prevented the matching.