BUGFIX: Missing metabolites now visible in MetExploreViz
BUGFIX: Metabolites that should be displayed were marked with hidden and were therefore not visible. Only happened if ' removeUnselectedGroups' = False.

NEW: Phases now have the variable objective_factor which can be used to influence the objective_functions of individual phases. The variable objective_factor acts as a multiplier for the individual objective_functions before all of them are summed up to get the final objective_function.