Implementation of a customized parsimonious FBA (#53)
NEW: two new functions implemented by a customized version of the pFBA. This one ignores linker identified by '_L_'. All other reactions are minimized, and the previous objective is set as constraint.

NEW: Tests for the two new functions.

NEW: Phases now have the variable objective_factor which can be used to influence the objective_functions of individual phases. The variable objective_factor acts as a multiplier for the individual objective_functions before all of them are summed up to get the final objective_function.

UPDATED: Metabolites that are not associated with any reaction are now also reliably copied.

BUGFIX: Missing metabolites now visible in MetExploreViz

BUGFIX: Metabolites that should be displayed were marked with hidden and were therefore not visible. Only happened if ' removeUnselectedGroups' = False.